renamed to os_linux_impl.hpp
