+++ src/hotspot/os/linux/cgroupSubsystem_linux.cpp
@@ -30,7 +30,7 @@
 #include "cgroupV2Subsystem_linux.hpp"
 #include "logging/log.hpp"
 #include "memory/allocation.hpp"
-#include "os_linux.impl.hpp"
+#include "os_linux_impl.hpp"
 #include "runtime/globals.hpp"
 #include "runtime/os.hpp"
 #include "utilities/globalDefinitions.hpp"
