+++ src/hotspot/os/linux/gc/z/zNUMA_linux.cpp
@@ -25,7 +25,7 @@
 #include "gc/z/zErrno.hpp"
 #include "gc/z/zNUMA.hpp"
 #include "gc/z/zSyscall_linux.hpp"
-#include "os_linux.impl.hpp"
+#include "os_linux_impl.hpp"
 #include "runtime/globals.hpp"
 #include "runtime/os.hpp"
 #include "utilities/debug.hpp"
