+++ src/hotspot/os/linux/osContainer_linux.cpp
@@ -28,7 +28,7 @@
 #include "runtime/globals.hpp"
 #include "runtime/os.hpp"
 #include "logging/log.hpp"
-#include "os_linux.impl.hpp"
+#include "os_linux_impl.hpp"
 #include "osContainer_linux.hpp"
 #include "cgroupSubsystem_linux.hpp"
