+++ src/hotspot/os/linux/trimCHeapDCmd.cpp
@@ -25,7 +25,7 @@
 
 #include "precompiled.hpp"
 #include "logging/log.hpp"
-#include "os_linux.impl.hpp"
+#include "os_linux_impl.hpp"
 #include "runtime/os.hpp"
 #include "utilities/debug.hpp"
 #include "utilities/ostream.hpp"
