+++ src/hotspot/os_cpu/linux_aarch64/os_linux_aarch64.cpp
@@ -33,7 +33,7 @@
 #include "code/nativeInst.hpp"
 #include "interpreter/interpreter.hpp"
 #include "memory/allocation.inline.hpp"
-#include "os_linux.impl.hpp"
+#include "os_linux_impl.hpp"
 #include "os_share_linux.hpp"
 #include "prims/jniFastGetField.hpp"
 #include "prims/jvm_misc.hpp"
