+++ src/hotspot/os_cpu/linux_arm/os_linux_arm.cpp
@@ -31,7 +31,7 @@
 #include "interpreter/interpreter.hpp"
 #include "memory/allocation.inline.hpp"
 #include "nativeInst_arm.hpp"
-#include "os_linux.impl.hpp"
+#include "os_linux_impl.hpp"
 #include "os_share_linux.hpp"
 #include "prims/jniFastGetField.hpp"
 #include "prims/jvm_misc.hpp"
