+++ src/hotspot/os_cpu/linux_ppc/os_linux_ppc.cpp
@@ -34,7 +34,7 @@
 #include "interpreter/interpreter.hpp"
 #include "memory/allocation.inline.hpp"
 #include "nativeInst_ppc.hpp"
-#include "os_linux.impl.hpp"
+#include "os_linux_impl.hpp"
 #include "os_share_linux.hpp"
 #include "prims/jniFastGetField.hpp"
 #include "prims/jvm_misc.hpp"
