+++ src/hotspot/os_cpu/linux_s390/os_linux_s390.cpp
@@ -36,7 +36,7 @@
 #include "interpreter/interpreter.hpp"
 #include "memory/allocation.inline.hpp"
 #include "nativeInst_s390.hpp"
-#include "os_linux.impl.hpp"
+#include "os_linux_impl.hpp"
 #include "os_share_linux.hpp"
 #include "prims/jniFastGetField.hpp"
 #include "prims/jvm_misc.hpp"
