+++ src/hotspot/os_cpu/linux_x86/os_linux_x86.cpp
@@ -32,7 +32,7 @@
 #include "interpreter/interpreter.hpp"
 #include "logging/log.hpp"
 #include "memory/allocation.inline.hpp"
-#include "os_linux.impl.hpp"
+#include "os_linux_impl.hpp"
 #include "os_share_linux.hpp"
 #include "prims/jniFastGetField.hpp"
 #include "prims/jvm_misc.hpp"
